revise authentication logic

+++ app/controllers/sessions_controller.rb
@@ -15,7 +15,9 @@ class SessionsController < ApplicationController
   end
 
   def destroy
-    session.delete(:user_id)
+    if current_user
+      session.delete(:user_id)
+    end
     redirect_to login_path
   end
 end
